+++ klum-ast-annotations/src/main/java/com/blackbuild/klum/ast/util/layer3/annotations/Cluster.java
@@ -23,21 +23,26 @@
  */
 package com.blackbuild.klum.ast.util.layer3.annotations;
 
+import com.blackbuild.klum.cast.KlumCastValidated;
+import com.blackbuild.klum.cast.checks.*;
 import org.codehaus.groovy.transform.GroovyASTTransformationClass;
 
 import java.lang.annotation.*;
+import java.util.Map;
 
 /**
- * Automatically implements the annotated method by providing access to all matching fields,
+ * Automatically converts the annotated field into a getter method providing access to all matching fields,
  * possibly filtered by the given annotation.
  *
- * <p>This is usually use to provide the API layer of a three layer model.</p>
+ * <p>if placed on a (potentially abstract) method, that method is replaced with such a getter.</p>
+ *
+ * <p>This is usually used to provide the API layer of a three layer model.</p>
  *
  * <p>For example:</p>
  *
  * <pre><code>
  * abstract class Database {
- *     {@literal @}Cluster abstract {@literal Map<String, User>} getUsers
+ *     {@literal @}Cluster {@literal Map<String, User>} users
  * }
  *
  * class MyApplicationDatabase extends Database {
@@ -61,8 +66,8 @@ import java.lang.annotation.*;
  *
  * <pre><code>
  * abstract class Database {
- *     {@literal @}Cluster abstract {@literal Map<String, User>} getUsers
- *     {@literal @}Cluster(Required) abstract {@literal Map<String, User>} getRequiredUsers
+ *     {@literal @}Cluster {@literal Map<String, User>} users
+ *     {@literal @}Cluster(Required) {@literal Map<String, User>} requiredUsers
  * }
  *
  * class MyApplicationDatabase extends Database {
@@ -75,33 +80,64 @@ import java.lang.annotation.*;
  * <p>In that example, <code>getUsers()</code> still returns all user fields, while
  * <code>getRequiredUsers()</code> only returns 'ddl' and 'dml'.</p>
  *
+ * <h2>Cluster Factories</h2>
+ *
+ * <p>In addition to the getter, a cluster factory is created (much like a collection factory), containing only the dsl methods
+ * of the respective cluster field</p>
+ *
+ * So with the above example, the following is correct:
+ *
+ * <pre><code>
+ * MyApplicationDatabase.Create.With {
+ *     users {
+ *         ddl {...}
+ *         dml {...}
+ *         monitoring {...}
+ *     }
+ * }
+ * </code></pre>
+ *
+ * <p>Using the {@link Cluster#bounded()} attribute, the setter methods can be restricted to be only available inside a factory
+ * (note that this is done by making the methods protected, not completely removing them).</p>
+ *
  * <p>The annotation can also be used for methods return Maps of Collections. If the (subclass of) Collection
  * is generic, the result is a map of the matching Collections. If the Collection does not use a generic parameter,
  * all collections would be returned.</p>
  *
- * <p>If the annotation is placed on a field, that field is transparently converted into a matching getter.</p>
- *
+ * <p>The annotation can also be set on classes or packages. If so, the bounded member is set for all cluster fields of a
+ * a class/all classes of its package. Other members are not allowed on classes/packages.</p>
  */
-@Target({ElementType.METHOD, ElementType.FIELD})
+@Target({ElementType.METHOD, ElementType.FIELD, ElementType.TYPE, ElementType.PACKAGE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @GroovyASTTransformationClass({
         "com.blackbuild.klum.ast.util.layer3.ClusterFieldTransformation",
         "com.blackbuild.klum.ast.util.layer3.ClusterTransformation"
         })
+@KlumCastValidated
+@NeedsType(Map.class)
+@NeedsReturnType(Map.class)
+@NeedsGenerics
+@NeedsOneOf(value = "bounded", whenOn = {ElementType.TYPE, ElementType.PACKAGE})
 public @interface Cluster {
 
     /**
      * If set, filters the results by the given annotation.
      * @return The annotation to filter on.
      */
-    Class<? extends Annotation> value() default Undefined.class;
+    @NotOn({ElementType.TYPE, ElementType.PACKAGE}) Class<? extends Annotation> value() default Undefined.class;
 
     /**
      * If set to false, null values are not included in the result.
      * @return To return or ignore null values.
      */
-    boolean includeNulls() default true;
+    @NotOn({ElementType.TYPE, ElementType.PACKAGE}) boolean includeNulls() default true;
+
+    /**
+     * If set to true, the setter methods for matching fields are only created inside a named factory.
+     * @return Whether the setter methods are only created inside a factory.
+     */
+    boolean bounded() default false;
 
     @interface Undefined {}
 }